feat(briefing): extract airport pipeline (commit 5/11)
Workstream 6 Step 1 — extracts the airport pipeline from briefing-service.js
into a focused module. Plan-conforming per resume-plan §3 commit 5; no drift
signals or design changes.

NEW: server/lib/briefing/pipelines/airport.js
  - discoverAirport({ snapshot, snapshotId }) — pipeline contract entry,
    returns { airport_conditions, reason }. Owns its writeSectionAndNotify
    and errorMarker .catch wrapper.
  - fetchAirportConditions({ snapshot }) — Gemini + Google Search live path,
    PRIVATE (no caller in server/, client/, or shared/ imports it directly;
    consolidator.js:652 references the function name only in a code COMMENT
    about data shape).
  - extractAirportJson — manual JSON-extraction fallback when safeJsonParse
    fails on markdown-wrapped Gemini responses, PRIVATE.
  - Bonus: moved the fetchAirportConditions JSDoc to its correct location
    (was misplaced above extractAirportJson in briefing-service.js).

briefing-service.js delta (2,907 → 2,750, net -157 lines):
  - import { discoverAirport } from pipelines/airport.js (no re-export — internal)
  - orchestrator airportPromise: replaced .then/.catch chain with direct
    discoverAirport() call (~7-line shrink, pipeline owns its own SSE write
    and errorMarker)
  - airportCount calc: airportResult?.airports?.length →
    airportResult?.airport_conditions?.airports?.length (line 2402)
  - final-assembly: airportResult || {fallback} →
    airportResult?.airport_conditions || {fallback} (line 2458)
  - deleted ~156 lines of airport code (1 helper + 1 main function + JSDocs)

Verification:
  - npm run typecheck (tsc --noEmit): clean
  - npm run lint (eslint client/src): clean
  - node --check on briefing-service.js, pipelines/airport.js: pass
  - Dynamic import smoke: airport.js exports { discoverAirport }
  - Grep audits: zero orphan local definitions, zero literal channel strings,
    zero external imports of fetchAirportConditions (confirms internal-only)
  - Both error pathways preserved:
    Pathway A (thrown) → Promise.allSettled captures → failedReasons.airport
    populated → fallback at lines 2461-2464 fires
    Pathway B (graceful) → fetchAirportConditions returns object with reason
    → discoverAirport wraps to { airport_conditions, reason } → orchestrator
    reads airportResult.airport_conditions at line 2458

Notable: airport's fetchAirportConditions never throws on graceful path
(internal try/catch returns fallback). The discoverAirport try/catch is
defensive against unexpected sync/import errors. Defense-in-depth `||`
short-circuit in discoverAirport preserves exact parity with the orchestrator's
prior `r || {fallback}` SSE-write fallback.

Per PLAN_workstream6_remaining-2026-05-02.md §2.1 (wider contract),
§2.3 (sed for line-range deletes).
